세세한 수정2
윈폼을 닫을 때 파이썬도 같이 끝나게 완성
신경망 갯수 더 증가
reward 세분화 (double형으로)
시간마다 reward에 -값 (움직이지 않는 행동에 패널티 부과)
전반적인 의사결정 랜덤성 감소

+++ CarDrive_1/Car.cs
@@ -48,7 +48,7 @@ namespace CarDrive_1
         const double turn = 3;
         Line[] lines = new Line[5];
         double[] distances = new double[5];
-        public int reward = 0;
+        public double reward = 0;
         public bool done = false;
         Line left = new Line();
         Line right = new Line();
@@ -75,9 +75,9 @@ namespace CarDrive_1
         public double[] getdistances() { return distances; }
         public double getdegree() { return degree; }
         public double getv() { return velocity; }
-        public int getreward()
+        public double getreward()
         {
-            int r = reward;
+            double r = reward;
             reward = 0;
             return r;
         }
